code cleanup and working serialization
