+++ Utilities/Settings.cs
@@ -1,28 +1,31 @@
-﻿using ActivityPub.Types;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Text.Json;
 
-namespace ActivityPub {
+namespace ActivityPub.Utilities {
+
+  /// <summary>
+  /// Static settings for ActivityPub
+  /// </summary>
   public static class Settings {
 
     static Settings() {
+      // Default settings with the two built in converters
       EntitySerializationOptions = new JsonSerializerOptions {
         Converters = {
-          new Entity.JsonConverter(),
-          new Link.JsonConverter()
+          new Json.Entity.Converter(),
+          new Json.Link.Converter()
         },
         WriteIndented = true
       };
 
+      ///Collect all default entity types. Used for serialization and deserialization:
       IEnumerable<System.Type> entityTypes = AppDomain.CurrentDomain.GetAssemblies()
         .SelectMany(s => s.GetTypes())
         .Where(type => typeof(Entity).IsAssignableFrom(type) 
           && type.IsClass && !type.IsAbstract);
 
-      ///Collect all default entity types:
       entityTypes.ToList().ForEach(type
         => _defaultTypes[type] 
           = (type.GetConstructor(
@@ -34,7 +37,7 @@ namespace ActivityPub {
               ?? throw new NotImplementedException(
                 "ActivityPub Entity Subtype: " +
                 type.FullName +
-                " does not have a private or public parameterless ctor"
+                ", does NOT have the required Private or Public Parameterless Constructor"
               )
             ).DefaultTypes
       );
@@ -67,12 +70,5 @@ namespace ActivityPub {
     static Dictionary<Type, IEnumerable<string>> _defaultTypes {
       get;
     } = new Dictionary<Type, IEnumerable<string>>();
-
-    public static void Test() {
-      var item = new ActivityPub.Types.Object {
-        Type = "message",
-        At = "User",
-      };
-    }
   }
 }
